refactor(dmr): split model discovery into openai-free dmrmodels package
pkg/runtime wired dmr.ListModels unconditionally, dragging the whole
openai-go SDK (via the DMR provider client and oaistream) into every
embedder that links the runtime, even Anthropic-only ones.

Move endpoint resolution and /models listing - which only use net/http -
into pkg/model/provider/dmr/dmrmodels and point the runtime at it.
dmr.ListModels and dmr.ErrNotInstalled remain as thin forwarders so
existing callers keep compiling and errors.Is keeps matching.
